+++ tibiawikisql/parsers/__init__.py
@@ -1,3 +1,29 @@
+#  Copyright (c) 2025.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#  http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
 from tibiawikisql.parsers.base import BaseParser
 
 from tibiawikisql.parsers.achievement import AchievementParser
@@ -11,3 +37,4 @@ from tibiawikisql.parsers.npc import NpcParser
 from tibiawikisql.parsers.imbuement import ImbuementParser
 from tibiawikisql.parsers.quest import QuestParser
 from tibiawikisql.parsers.house import HouseParser
+from tibiawikisql.parsers.outfit import OutfitParser